it's just getting worse
