+++ include/aws/s3/private/s3_meta_request_impl.h
@@ -176,10 +176,6 @@ struct aws_s3_meta_request {
     aws_s3_meta_request_telemetry_fn *telemetry_callback;
     aws_s3_meta_request_upload_review_fn *upload_review_callback;
 
-    /* Customer specified callbacks to be called by our specialized callback to calculate the response checksum. */
-    aws_s3_meta_request_headers_callback_fn *headers_user_callback_after_checksum;
-    aws_s3_meta_request_finish_fn *finish_user_callback_after_checksum;
-
     enum aws_s3_meta_request_type type;
     struct aws_string *s3express_session_host;
